Update HF GPT-2 test to forward-backward
